+++ hivez/lib/src/builders/builders.dart
@@ -1,0 +1,7 @@
+library;
+
+import 'package:hive_ce/hive.dart';
+import 'package:hivez/src/boxes/boxes.dart';
+
+part 'box_creator.dart';
+part 'box_config.dart';